# src/Bank.Digital.OrderService.Api/swaggerV1.json
{"swagger":"2.0","info":{"version":"1.0","title":"OrderService API"},"paths":{"/api/v1.0/Orders":{"get":{"tags":["Orders"],"summary":"получаем все созданные заявки данным пользователем","operationId":"GetOrderList","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"applicantId","in":"query","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Order"}}}}},"post":{"tags":["Orders"],"summary":"создание новой заявки","operationId":"CreateOrder","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"request","in":"body","required":false,"schema":{"$ref":"#/definitions/OrderCreateRequest"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Order"}},"400":{"description":"Bad Request"}}}},"/api/v1.0/Orders/{id}":{"get":{"tags":["Orders"],"summary":"получаем полную информацию по конкретной заявке","operationId":"GetOrder","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","format":"uuid"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Order"}},"404":{"description":"Not Found"},"400":{"description":"Bad Request"}}}},"/api/v1.0/Tests/hadoop":{"get":{"tags":["Tests"],"operationId":"Get","consumes":[],"produces":[],"parameters":[{"name":"filename","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}},"/api/v1.0/Tests/exception":{"post":{"tags":["Tests"],"operationId":"Exception","consumes":[],"produces":[],"parameters":[{"name":"name","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Success"}}}}},"definitions":{"Order":{"type":"object","properties":{"id":{"format":"uuid","description":"ID заявки","type":"string"},"applicantId":{"format":"int32","description":"Аппликант/проспект/ человек желающий оформить услугу","type":"integer"},"state":{"description":"Состояние заявки","enum":["New","Filling","Revision","Rejected"],"type":"string"},"dateAdd":{"format":"date-time","description":"Дата создания заявки","type":"string"},"productId":{"format":"int32","description":"ID открываемого продукта ( РКО, гарантии, кредит)","type":"integer"}}},"OrderCreateRequest":{"required":["applicantId","dateAdd","productId"],"type":"object","properties":{"applicantId":{"format":"int32","description":"Аппликант/проспект/ человек желающий оформить услугу","maximum":2147483647.0,"minimum":1.0,"type":"integer"},"dateAdd":{"format":"date-time","description":"Дата создания заявки","type":"string"},"productId":{"format":"int32","description":"ID открываемого продукта ( РКО, гарантии, кредит)","maximum":2147483647.0,"minimum":1.0,"type":"integer"},"ip":{"description":"IP адрес с которого была создана заявка","type":"string"},"utm":{"description":"UTM метка","type":"string"},"googleAnalyticsId":{"description":"аналитика от гугла","type":"string"},"region":{"description":"город","type":"string"},"browser":{"description":"браузер клиента","type":"string"},"browserUserAgent":{"description":"юзер агент браузера","type":"string"}}}}}